requirements.txt
adblock==0.6.0
aiodns>=2.0.0
aiofiles==0.8.0
aiohttp==3.7.4
aiohttp-socks==0.7.1
aiohttp-basicauth==1.0.0
aioipfs==0.5.8
aiojobs==0.3.0
aiorwlock==1.0.0
aiosparql==0.12.0
aiosqlite<0.17.0,>=0.16.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncio-throttle==1.0.2
attrs==22.1.0
base58>=1.0.2,<2.0
bcp47==0.0.4
bencodepy==0.9.5
bitarray>=1.6.1
cachetools==5.2.0
claptcha==0.3.3
colorhash==1.1.0
colour==0.1.5
feedgen==0.9.0
feedparser==6.0.10
frozendict==1.2
filelock==3.8.0
gitpython==3.1.27
html2text==2020.1.16
ignition-gemini==0.1.11
importlib-metadata==4.6.1
ipfshttpclient==0.7.0
jinja2==3.0.1
json-traverse==0.4
jsonschema==3.2.0
jwcrypto==1.3.1
logbook==1.5.3
markdown==3.3.7
memory-profiler==0.60.0
mkdocs==1.3.1
mkdocs-bootswatch==1.1
mode==4.4.0
multiaddr>=0.0.8
omegaconf==2.1.1
orjson>=3
otsrdflib==0.5.0
pillow==9.0.1
protobuf==3.6.1
psutil==5.9.2
ptpython==3.0.20
pydotplus==2.0.2
py-multicodec==0.2.1
py-multihash==2.0.1
pycryptodomex==3.9.8
pygments==2.13.0
pyld==2.0.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-magic==0.4.27
python3-crdt>=1.0.3
py-multibase==1.0.3
pyyaml==6.0
pyparsing==2.4.7
pyzbar==0.1.9
qrcode==7.3.1
qasync==0.22.0
qtawesome==1.1.1
random-username==1.0.2
rdfextras==0.4
rdflib-pyld-compat==0.1.0
rdflib-sqlalchemy==0.5.3
rule-engine==3.5.0
SPARQLWrapper==1.8.5
strict-rfc3339==0.7
tortoise-orm==0.17.6
urnparse==0.2.0
validators==0.20.0
yarl==1.5.1
aiogeminipfs @ https://gitlab.com/galacteek/aiogeminipfs/-/releases/0.2.1/downloads/aiogeminipfs-0.2.1-py3-none-any.whl#sha256=4bb42f2c0f2f74a0e7bc35a1df37d63763b573d8da5d900918f9829a9d782ee7
aiopubsub @ https://gitlab.com/api/v4/projects/40071322/packages/generic/aiopubsub/3.1.0/aiopubsub-3.1.0-py3-none-any.whl#sha256=520d3ec43f470fd612320e782e4bb333cded7b7b8cc3d876cfe7b4bc79a2a19e
rdflib @ https://gitlab.com/galacteek/rdflib/-/releases/5.0.0/downloads/rdflib-5.0.0-py3-none-any.whl#sha256=ddcae9e193d999d810bafaeea9fcf55ec7024486fec51053e9c8eef63e4d179f
rdflib-jsonld @ https://gitlab.com/galacteek/rdflib-jsonld/-/releases/0.6.0-dev/downloads/rdflib_jsonld-0.6.0.dev0-py2.py3-none-any.whl#sha256=d2e3ba9fd3372ecc9321ca5a66da58b84d21205c67554bbb75154f033f12b8d0
SparQL-Burger @ https://gitlab.com/galacteek/SPARQL-Burger/-/releases/1.0.2/downloads/SparQL_Burger-1.0.2-py3-none-any.whl#sha256=3e6932ee1f69d434e4691ac99797d34dce148b720f0aa0cf2f3e84b5fae729bd
